pkg_install: Disable check_requires().
Still has a number of problems, and the very last thing we want is to be
stuck in a situation that we might not be able to recover from.  The bug
this code was implemented to solve does not really apply to us, and our
users are trained to use pkgin not pkg_add.